Add support of Statiq virtual directories
When testing a Statiq site without a CNAME configured, it is typically
run out of a virtual directory, and testing need to be done out of the
same named virtual directory.

This PR adds the ability to configure the name of this virtual directory
via a build parameter, which can be passed into the SetParameters method
in the recipe.cake file.