(#8) Pin to earlier version of Cake.Git
There is "something" not quite right with the version 2.0.0 of this
addin, which means that it isn't possible to commit changed files to a
repiository.  This has the effect of meaning that we can't publish the
documentation changes.

Pinning to the earlier version means that things "just" work, so we will
go with that just now.